# my-express-api/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UsersResponse": {
				"properties": {
					"isVerified": {
						"type": "boolean"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"isVerified",
					"email",
					"name"
				],
				"type": "object"
			},
			"DefaultSelection_Prisma._36_UserPayload_": {
				"properties": {
					"isVerified": {
						"type": "boolean"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"isVerified",
					"password",
					"name",
					"email",
					"id"
				],
				"type": "object"
			},
			"User": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_UserPayload_",
				"description": "Model User"
			},
			"UserRegistration": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email",
					"name"
				],
				"type": "object"
			},
			"UserLoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object"
			},
			"UserLogin": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"DefaultSelection_Prisma._36_LineItemPayload_": {
				"properties": {
					"budgetId": {
						"type": "string"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"budgetId",
					"categoryId",
					"amount",
					"name",
					"id"
				],
				"type": "object"
			},
			"LineItem": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_LineItemPayload_",
				"description": "Model LineItem"
			},
			"DefaultSelection_Prisma._36_CategoryPayload_": {
				"properties": {
					"colorHex": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"colorHex",
					"name",
					"id"
				],
				"type": "object"
			},
			"Category": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_CategoryPayload_",
				"description": "Model Category"
			},
			"DefaultSelection_Prisma._36_BudgetPayload_": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"savings": {
						"type": "number",
						"format": "double"
					},
					"startingCapital": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"savings",
					"startingCapital",
					"name",
					"id"
				],
				"type": "object"
			},
			"Budget": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_BudgetPayload_",
				"description": "Model Budget"
			},
			"BudgetWithLineItems": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"startingCapital": {
						"type": "number",
						"format": "double"
					},
					"savings": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "string"
					},
					"lineItems": {
						"items": {
							"$ref": "#/components/schemas/LineItem"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"startingCapital",
					"savings",
					"userId",
					"lineItems"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BudgetResponse": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BudgetWithLineItems"
					}
				],
				"nullable": true
			},
			"BudgetCreate": {
				"properties": {
					"name": {
						"type": "string"
					},
					"startingCapital": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"startingCapital"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BudgetCreateBody": {
				"$ref": "#/components/schemas/BudgetCreate"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "my-express-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UsersResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRegistration"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLogin"
							}
						}
					}
				}
			}
		},
		"/lineitems/{lineItemId}": {
			"get": {
				"operationId": "GetLineItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/LineItem"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Lineitems"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "lineItemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateLineItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LineItem"
								}
							}
						}
					}
				},
				"tags": [
					"Lineitems"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "lineItemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LineItem"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteLineItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LineItem"
								}
							}
						}
					}
				},
				"tags": [
					"Lineitems"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "lineItemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lineitems/{budgetId}": {
			"post": {
				"operationId": "CreateLineItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LineItem"
								}
							}
						}
					}
				},
				"tags": [
					"Lineitems"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "budgetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LineItem"
							}
						}
					}
				}
			}
		},
		"/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Category"
							}
						}
					}
				}
			}
		},
		"/categories/{categoryId}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Category"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Category"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/budgets": {
			"get": {
				"operationId": "GetBudgets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Budget"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Budgets"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBudget",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BudgetResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Budgets"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BudgetCreateBody"
							}
						}
					}
				}
			}
		},
		"/budgets/{budgetId}": {
			"get": {
				"operationId": "GetBudget",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BudgetResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Budgets"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "budgetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBudget",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Budget"
								}
							}
						}
					}
				},
				"tags": [
					"Budgets"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "budgetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Budget"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBudget",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Budget"
								}
							}
						}
					}
				},
				"tags": [
					"Budgets"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "budgetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	],
	"tags": [
		{
			"name": "Users",
			"description": "Manage user data aswell as registration and login"
		},
		{
			"name": "Budgets",
			"description": "Manage budgets created by the user"
		},
		{
			"name": "Lineitems",
			"description": "Manage lineitems related to a budget"
		},
		{
			"name": "Categories",
			"description": "Manage categories related to lineitems"
		}
	]
}